rename store state
